arch-vega: Pass s_memtime through smem pipe
The Vega ISA's s_memtime instruction is used to obtain a cycle value from the GPU.
Previously, this was implemented to obtain the cycle count when the memtime instruction reached the execute stage of the GPU pipeline.
However, from microbenchmarking we have found that this under reports the latency for memtime instructions relative to real hardware.
Thus, we changed its behavior to go through the scalar memory pipeline and obtain a latency value from the the SQC (L1 I$).
This mirrors the suggestion of the AMD Vega ISA manual that s_memtime should be treated like a s_load_dwordx2.

The default latency was set based on microbenchmarking.

Change-Id: I97040b8448c96d56377d279961717d6978d5e03b